Republish dependent editions on contact update
This came through as a Zendesk issue. A user had updated
a contact on an org page, but an editions using that
contact (as markdown) did not update.

By investigating the callback, it seems plausible that
the async calls to republish may execute before the db
transaction of updating the contact and its
relationships has finished. So the calls are being made,
but on stale data.

I was not able to reproduce this, and the existing tests,
which run Sidekiq inline, seem to send the correct load
to pub api. Nonetheless, it seems a safe bet to replace
the callback in order to prevent further issues. Will
restrict the callback to only run on update. `after_commit`
has some transaction rollback implications, meaning we lose
the guarantee of rolling back the update.