Override `read_attribute_for_validation`
By default, Rails expects all errors to be related to readabled
attributes within a model. As `details` is a hash, we override the
default `read_attribute_for_validation` to check the attribute we’re
attempting to read begins with a `details_` prefix. If it does, we
attempt to read it from the hash, otherwise we bubble up to the original
implementation.