Content Block Documents store a reference to the latest edition
Whitehall main use these column names. Right now we are auto publishing to live so our latest and live will be the same. This may not always be the case so we expect their values to be duplicate to start. We may remove one of them if turns out we never need them both.