Add `creator` as first Content Block Edition Author
When a CB Edition is created it should have a
`creator`, and that creator should be the first in
it's list of `edition_authors`. This copies
behaviour of existing Edition in main app.

We will shortly extract this behaviour into
concerns.

In the near future, when we have edit/update
functionality, the list of `edition_authors` will
be added to on `save`.