Add a method to get a schema for an edition
This allows us to get the edition’s schema easily for use in validation.
I’ve added this to a concern that we’ll build on in future commits. We
can also override the schema with an `attr_writer`. This is preferable
to a mock, as mocking every time we create an object is problematic, as
they’re difficult to clean up afterwards, and have unintended effects.