Rename Content Object Store Identifiable concern for clarity
Following a discussion[1] we want to clarify what we expect this concern to contain.

We want this to be more than identification, we want it to be everything that connects an edition to a document.

This does diverge us from the main Whitehall we were referencing which may make this harder to merge back together later. But for now the short term benefit of the team knowing how to use it clearly takes priority.

[1] https://github.com/alphagov/whitehall/pull/9271#pullrequestreview-2180611132