ContentBlockEditions accept ContentBlockDocuments as nested attributes
It was becomign cumbersome to remember which attributes were for the document and which were for the edition when it came to creating and updating.

Instead of having to track this we can use the Rails `accepts_nested_attributes_for` to help us take care of this. The main Whitehall repository takes this approach[1].

With it we’re able to send around a single blob of params, pass them to the Edition ORM and it’ll figure out what to do, including validations.

* We have to make one override to the title error message so we see “Title can’t be blank” rather than “Content document title can’t be blank”.
* We also adjust the hidden field for block tag so it puts the params in the same place. Inside `content_block_edition`, rather than `content_block_document_attributes ` at the top level. We’ll now have `content_block_edition: { details: {}, content_block_documentattributes: {}}`.
* In renaming these attributes we also fix the form so the error messages correctly link to the title field on error

[1] https://github.com/alphagov/whitehall/blob/7cbd2b9b6d97df9741befcc4700fb573fb2e1960/app/models/edition.rb#L82]