New ContentBlockDocuments know their latest edition
When tackling the updating work I needed to find a way to tell the Document which one of it’s many Editions was the latest one.

Referring back to Whitehall main they use live_edition_id and latest_edition_id. We might not need both but we align for now until we have more information.

Though Edition accepts nested parameters for a Document the `edition.id` needs to first exist as the result of a creation, so we reluctantly add this as another database operation.