ContentBlock Editions can be updated
* The edit form view is not reusing the existing _form partial which is unconventional. It’s needed because of the way we access the fields for a new record is different from editing.

Usually we’d use a new active record object that was initialised with the fields we want. For creation we need build the form around the loaded ContentBlockSchema properties so we can present fields not in our conventional schema.

For update we can’t build the form by accessing the schema in the same way as the __values__ we’re editing now conventionally come from the persisted object we’re working on

* Next we’ll refactor the controller logic into an Update service, inline with the Creation pattern