add simple HasAuditTrail concern
we'd like to have an Audit Trail for our Content
Blocks, so that we can see the history of who
created/updated them.

Following existing pattern in Whitehall, this adds
a slimmed down version of the AuditTrail concern.
I've tweaked the naming to be more readable/fit
with our current pattern.

This just ensures a ContentBlockVersion is created
when an Edition is created.