Updating content blocks renders validation errors
* Instead of rescuing from `ActiveRecord::RecordInvalid` in the controller (as we do with create) we have to do more work to support updates.

We start with a traditional update where the user is changing the form fields of an existing record. However, when this goes through we’ll create a new edition and update the document.

If this goes well we return them to the new edition.

If this goes badly we need to render back to edit but with form errors that are attached to the original object. Plus we need to render show and cancel links which require an object with an ID. Both of these require us to do a check on the original object for validity (manually attaching the document ID so it can check that too) and then if OK we proceed to create a new record.

* We have to be careful that this Edition update doesn’t interpret the nested params and create a new document.

This more concise line fails the tests:

```ruby
ContentObjectStore::ContentBlockEdition.create!(edition_params.merge(content_block_document_id: original_document_id))
```

and so we have to update the document separately and assign the relationship before we save:

```
new_content_block_edition = ContentObjectStore::ContentBlockEdition.new(edition_params)
new_content_block_edition.content_block_document_id = @original_content_block_edition.document.id
new_content_block_edition.save!
```