Add feature flag for Content Object Store
while we are still developing the object store we
do not want it released to the public.

By default this will be turned on in development,
but off in other environments until we’re ready to
test/roll out.

We will shortly use it in a Content Object Store controller

Feature flag docs: https://github.com/alphagov/whitehall/blob/main/docs/feature_flags.md