add Content Block Document timeline
This is a component to construct a timeline for a
Document based on the versions of it's editions.
(which can be found by calling
`content_block_versions` on a Document)

We currently only have 'created' events for our
Editions - once an edition is created it is
published straight away, and so we don't have
drafts/schedules/edits. As our workflow becomes
more complex we can add events and change how the
data is presented.