Content block edition factory allows assigning existing document
When testing the `document.latest_edition` behaviour the factory couldn’t easily support the creation of an Edition with an existing Document. It could only create a new Edition and new Document.

The use of `content_block_document_attributes` was an attempt to simulate how the Edition could take parameters for the nested document object. This works for the creation of a first_edition but not for a second_edition:

```
content_block_document_attributes { FactoryBot.attributes_for(:content_block_document, :email_address) }
```

Instead of messing around with traits to try and support both:
```
    trait :skip_document_create do
      transient do
        skip_document_create { true }
        content_block_document_attributes { nil }
      end
    end
```

We switch to the more traditional usage of `create(edition, document: document)` when we want to associated with an existing document. Otherwise we use create(edition, :block_type) to create a new first_edition.