Allow users to navigate from the higher Document level
We want to allow users to navigate to the parent concept of content_block_documents, rather than going straight into the latest editions.

This will help us group timeline information at the Document level.

This is at odds with how Whitehall main has organising their user flow but after checking we learnt that they have had issues with having to work backwards to find the latest document on each page, and are curious to see if this pattern works better.

1. As the first step for this work, we duplicate the feature tests into a temporary directory in order to make sure the old user journey works, as well as the new one. We’ll remove this duplication later
2. We copy over the index and show views from Edition, we’ll keep both sets of views going forward