Fix 'search results resetting locale to en' bug
Navigating to, for example, "Features (Español)" and searching for
a document to feature, would submit and redirect to the search
results page on the "Features (English)" tab, since no locale was
set in the form submit URL. This meant it was impossible to search
for documents by their foreign language title.
This is fixed in this commit.

Note that there is still a question as to what the ideal behaviour
would be on the non-English search page, in terms of 1) the
language used to search (English or non-English?) and 2) the
language displayed in the search results. Current behaviour is
that it takes a non-English search term, but surfaces the result
'in English' (but then features the non-English version when you
choose to feature it). We're awaiting guidance from FCDO on the
behaviour they'd like ideally, but that shoudn't block this PR,
and can come as a later piece of work. This PR therefore adds a
test that documents the existing behaviour, and should provide a
good basis for a test if we decide to tweak the behaviour in
future.