Consistently keep the locale in the 'features' URL
Navigating to the "Features" tab for an Organisation takes you
to "/features.en" - i.e. the locale is in the URL, even though
Organisations are only ever in English. Submitting the form
on the "Features" tab (to search for a document to feature) then
redirects you to "/features" (without the locale). For consistency
we should always provide a locale for this path.