Legacy ContentBlockDocuments are updated with their latest edition ids
During https://github.com/alphagov/whitehall/pull/9305 we added in the association and then started using latest_edition in anger here https://github.com/alphagov/whitehall/pull/9307.

This works locally on a clean database.

Whilst we no real users are using this service, theoretically the team could have used the service on integration and on their own environments to create Documents. If those Documents were prior to the association then they’ll not have a `latest_edition` and so their views will break, as we found when pairing this afternoon.

It looks like there is support for a data migration (as opposed to a normal Rails migration) however I don’t think we fit any of the criteria:

```
* requires interacting with another system, such as publishing-api or the search
  index
* depends on specific data being present in the database - for example, a third
  party running our code won't have access to a copy of our production data
* is a long-running data change which you want the flexibility to be able to run
  during quiet times
```

Also we see another recent migration doing updates[1] which points to this being a conventional way of doing this within Whitehall.

[1] https://github.com/alphagov/whitehall/blob/main/db/migrate/20240723115422_rename_editionable_worldwide_organisation_polymorphic_relationships.rb