add `state` column to Editions
we would like to introduce an 'interstitial' page
between creating an Edition and publishing it, in
order to allow users to confirm the content they
are adding. This ties into the need for different
states of an Edition (submitted, draft, published
etc) - in Whitehall this is encapsulated in the
Workflow concern and controller.

here we add the `state` field to a CB Edition in
order for us to use it in workflow in following
commits.