Add task to bulk remove links to domains
Occasionally, entire domains which are regularly linked to on GOV.UK
break. For example, there are over 1000 links to storify.com which was:

> a social network service that let the user create stories or timelines
> using social media such as Twitter, Facebook and Instagram

(https://en.wikipedia.org/wiki/Storify)

... until it went out of business in 2018.

Sometimes these domains just disappear, but occasionally they're
hoovered up by some malicious third party and end up doing spammy
things, or, worse, phishy things. (Note: storify isn't an example of
that - it's simply down).

In these situations, we've got a choice with no clearly good options:

- Leave the links on GOV.UK pointing to malicious sites (not good)
- Attempt to contact the "owners" of the affected pages on GOV.UK and
  have them fix the links (not practical, since many pages older than a
  couple of years are effectively unowned)
- Mark up the links in some way to show that they're broken and warn
  users before they click on them (confusing / scary UX)
- Archive the content (only works if the content with the broken links
  can actually be archived, which isn't always true. Also it's never
  true at the moment, as there's no way to meaningfully "archive" content
  on GOV.UK yet).
- Replace the links with their bare link text (what this PR does)
  (potentially confusing in some situations, where the link text might
  be "click here" or similar).
- Redirect the links to a "this link is broken" page on GOV.UK (looks
  bad, potentially scary UX)
- Redirect the links to another site (only works if there's a sensible
  place to redirect to) (this will be addressed in a subsequent commit)
- Remove the links entirely, including their link text (would break
  sentences / potentially make content unreadable / inaccurate)

This commit adds a rake task which, given a domain, finds all the
content in Whitehall which links to that domain, and replaces the links
with the plain text link text.

I wanted to do this with the govspeak parser, instead of doing a regex
replace, but unfortunately it doesn't cleanly round trip (i.e. if you
parse a doc and then emit it again as govspeak, there will be lots of
differences, including but not limited to whitespace).

The regex approach does have some risk that we could incorrectly
replace something that's not a link (e.g. something inside a code block
might look like a link, but not actually be one). But I think the risk
is low, considering the subset of markdown that's permitted in govspeak.

I've tried to make the patterns as strict as possible to reduce the risk
of this being called with bad arguments and doing widespread damage. The
domain has to match the one in the link exactly, and we don't support
very short domains.