New task to delete an attachment for old superseded editions
There was an HTML attachment that was still live and still had state "deleted: false" in the database but the edition it is associated with is superseded, and the newest edition does not have any associations with the attachment so it can't be deleted. New rake task to remove these kinds of attachments.
Also moved attachment related tasks into an attachments folder for navigation