Revert "Disable republishing of embassies index page"
This reverts commit 38a978996c515a92cc466108068f21dd0bb14bd4.

From the original commit:

> Republishing this page was causing an incident because it was running
> some extremely slow database queries.

> We think a migration run as part of the editionable worldwide
> organisations cleanup work may have resulted in a loss of performance
> for queries run during the embassies index presenting process

In the subsequent investigation we did find a lot of N+1s in the
EmbassiesIndexPresenter.

We're also looking into the possibility that the Embassies Index was
actually being published more times than it needed to be (due to at
least some of the `after_commit` callbacks that this commit touched).