add Publish Edition Service
we would like to introduce an 'interstitial' page
between creating an Edition and publishing it, in
order to allow users to confirm the content they
are adding.

Instead of immediately publishing on creation of
an edition, we will be extracting the publishing
functionality to be done when the user confirms
they would like to publish.