add a 'draft -> publish' flow on creation of an Edition
we would like to introduce an 'interstitial' page
between creating an Edition and publishing it, in
order to allow users to confirm the content they
are adding. This ties into the need for different
states of an Edition (submitted, draft, published
etc) - in Whitehall this is encapsulated in the
Workflow concern and controller.

The `create` function now re-directs to a view to
show the user their content to confirm they would
like to publish.