Update the Embassies Index from WWOrg
Stop republishing the Embassies Index when WorldwideOffices and their
Contacts are edited. Instead, republish the Embassies Index when
a WorldwideOrganisation is updated to a "post publication state"
(published, superseded, withdrawn or unpublished).

We're currently re-presenting the Embassies Index to Publishing API
multiple times whenever a WorldwideOrganisation is edited, if poss we'd
like to only do that the one time.