Add task to bulk redirect links to domains
This is very similar to the previous commit, which introduced a task to replace links to a certain domain with the plain text link text.

For some domains, there's a URL somewhere else on GOV.UK or on the internet where we could re-point all of the links. For example, anything pointing at http://www.cesg.gov.uk/ (which is an NXDOMAIN) could be repointed at https://www.gov.uk/government/organisations/cesg

There's a fair bit of duplication between LinkRemover and LinkRedirector, but they're different enough that it's hard to reuse the code without making them hard to read. So I've just accepted the duplication.