Add a GetLinkedContentItems service
This allows us to call the `get_content_items` endpoint with the
required arguments, and marshal the response into the correct format.

Currently we try to get the `primary_publishing_organisation` from the
resulting links and then fetch the organisation if it exists. Going
forward, we should handle what to do if that link is missing and attempt
to glean what the associated organisation is from some other method
(such as the publishing app)