Add guard clause to prevent republishing of deleted WWorgs.
This prevents a condition where Users where users find themselves unable
to edit any roles associated with a deleted draft WWorg, with an any
attempt to simply throwing an error. The guard clause helps as Editions
are normally soft deleted, and the default scope of `Edition` filters out
any with a state of `deleted`.