Ensure topical events can be removed from organisations
The `delete_absent_topical_event_organisations` method was operating directly on the params object, but the `organisation_params` method that was passed to the organisation model was grabbing a fresh copy of the parameters, and therefore the `_destroy` parameter was not included.

By using the `organisation_params` method inside `delete_absent_topical_event_organisations`, we make sure that topical event organisation associations can be removed.

We also added a test to verify this behaviour.