Prevent users from enabling access limiting for documents that are not owned by their organisation
Rather than using conventional Rails validation, we decided to add this to the draft edition updater. We took this approach because the "validity" of the edition in this situation is entirely dependent on the request context (i.e. the user making the request), so placing the validation on the model layer itself didn't seem appropriate.

There are also situations where editions are updated in queue workers or by the scheduled publishing robot, so again performing this check on the model did not seem appropriate.

We also take advantage of the human-friendly failure reason messages provided by the edition updater to provide more useful feedback to the publisher.