Add rake task to report on deleted assets that are still live
This adds a rake task which produces a report listing all attachments
that are only attached to non-visible editions (i.e. the publisher
thinks they are deleted) but are not deleted (and therefore publicly
visible) in Asset Manager.

Also adds a `with_alternative_format_provider` trait to the publication
factory, since an alternative format provider is required before
attachments can be attached to a publication.