Make government a real association of editions
Prior to this commit, Whitehall inferred a relationship between editions and governments using the political flag (now superseded) and a date relating to the edition (usually first edition publishing date).

This meant that it wasn't possible for content editors to override the government in the event that a piece of content should actually relate to a different government. This is a common situation in the days after an election.

This commit makes the government a real association of a document edition, and allows GDS editors to set the value of the association in the edition form using a select element.

It also preserves the previous behaviour of setting the initial political status at the time of the publication of the first edition based on the appropriate date attribute for each edition. In this case we select the government that was in power at the time. Any edition with a government association is considered to be political.

Finally, we preserve the behaviour of sending a government link for each edition to Publishing API regardless of whether or not the edition is political. We use the default government for each edition to do that if no government is set.