Add data migration to assign government to political editions
This is part of the work to enable documents to be associated with a particular government.

There will be a period between the code being deployed and this data migration running where some political documents will not have the checkbox checked in the Whitehall interface because they will not have a government assigned to them.