Avoid duplicate form headings by using a page_title
Before this work introduced a new `hide_title` option which allowed us to stop showing the `title` of a page if one was set by a nested component, such as a radio form.

The radio component requires a heading.

Instead of tweaking the layout used by Whitehall proper we switch to using page_title to achieve the same.

Changing the layout felt like a reach. Though it works it would require every reader to understand what hide_title was, when to use it and whether they should use it. Given it was just for one of our pages I think this probably wasn’t worth the cost.