add multi-step logic to edit form controller
we want to streamline the logic for the
content block forms that are multi-page, so that
it's easy in the future to add new pages to the
user journey and/or move the pages around.

this shows how we might start to do this without using
an external library - we take a step parameter and execute
the logic needed to show that step.