Remove edition roles when edition has been deleted.
This fixes a bug where edition role associations remained in the database even after a worldwide organisation had been deleted. This caused issues when trying to look up the editions related to a role, because all edition queries are scoped by default to exclude deleted editions.

Since worldwide organisations became an edition deleting them has transitioned them to the deleted state instead of deleting them from the database. This means that the destroy dependancy option on the active record edition roles association no longer worked.

We've therefore moved the behaviour to tidy up the edition's role associations on to the `EditionDeleter` service.