Fix the showing of embedded content by changing how we fetch it
This old implementation worked locally but doesn’t work on Integration. Likely due to the size of the database. See full details in the Publishing API change here[1].

We temporarily lose pagination which we got for free with the old endpoint. We’ll have to work on adding that back in shortly.

We also make a direct call to the Publishing API using Net::HTTP. We
will look to implementing a new method in GDS Adapters once we're happy
this works.

[1] https://github.com/alphagov/publishing-api/pull/2846