Enable overriding of government associated with an edition
This is necessary because in some cases, e.g. on days where the government changes, content is associated with the wrong government. For example, a document might be associated with the previous government, but only published on the day after that government has left office.

We want to make sure that a government isn't deleted whilst there is still content associated with it, so we set the foreign key up to restrict deletion in that case. When we build an interface for removing governments we will ensure that the user is aware that there are documents associated with the government before they attempt to delete it.