Add an interface for GDS editors and admins to override the government associated with an edition
Decided to break this out into a view component so that it was possible to push some of the tests a little further down the pyramid.

Overriding the government will change the government displayed on the history mode banner on historic documents.

Annoyingly, because the conditional checkbox behaviour relies on Javascript, the best we can do to test that is working is assert that the government selection element is wrapped in a conditional div belonging to the political checkbox