Fix error caused by some edition types not implementing organisations association
The draft updater checks that the user is not about to limit their own access to a document.

However, not all documents can be access limited, or access is controlled by other means. Corporate information pages are an example of this.

This commit ensures that the updater only validates that the user is not about to limit their own access in contexts where the document is tagged to one or more organisations.