Get content that embeds a target via GDS API Adapters
As it happens the Net::HTTP idea didn’t work on Integration due to a connection issue.

This commit depends on the GDS API Adapter extension[1] which first adds this method.

Naming continues to be a battle. Coining a concise term that captures the concept and the relationships is hard. We have reusable editions that are embedded and linked to by ‘normal’ editions. It is true to say that these normal editions are dependent on the reusable ones but given we’ve got the endpoint named ending in `/embedded` I’ve leaned into that for now.

[1] https://github.com/alphagov/gds-api-adapters/pull/1283