move updating live/latest to publishable concern
and remove it from the service methods. I think
that the the publishable concern should handle
updating these fields, as they are directly
related to the publishing workflow, and the
services only responsible for handling content
generated by the forms.

we also ensure that publishing an Edition makes it
the latest and the live edition, but scheduling
only makes it the latest Edition. Because we do
not have a draft status, and we are only allowing
one scheduled edition at a time, when an edition
is published it should always be the latest
edition, unless it gets overwritten by a new
scheduled edition.