Prevent non-political documents having invalid value of political
The first iteration of the history mode controls component set the value of the political attribute on the edition to an empty string if the checkbox was not checked.

We now set the value to 0, as per the original implementation, and add some validation to prevent the issue being reintroduced.