allow audit trail for updates to Editions
This allows us to record an audit trail of when an
Edition is updated. Previously we only had events
to track when an Edition was created - because our
Content Blocks are published as soon as they are
created, this was sufficient, as an Edition would
never be 'updated', only superseded by a new
Edition.

Now we have introduced Scheduling, an Edition can
be updated. It will be created and set in the
`scheduled` state, and then `published` by the
scheduling worker.

For this reason we introdue the `record_update`
method, and create an 'updated' event with the
state - this will also be handy in the future
should we ever introduce new states (e.g.
withdrawn, draft etc.)

These versions provide an Edition-level database
log of events for us, but the user need on a
Document-level view is slightly different. After
reviewing the existing Whitehall language we
decided to remove the 'created' and 'changed'
events on the Timeline and only show the updates
when a state has changed - currently that should
capture any changes made, as when an Edition is
created it either scheduled or published at the
same time. [1]

[1] https://gds.slack.com/archives/C062Z1FA2ES/p1725375526406069?thread_ts=1725360904.799849&cid=C062Z1FA2ES