Dequeue any previous schedule
This commit is super simple in order to make it work. Next we'll look at
the design of the code and see if we can reorganise the service layer to
make this fit more neatly.

I ran into an unusual issue when testing this. It looks similar to this
unsolved issue in Rails[1]. When I try to use the
ActionDispatch::Routing::RoutesProxy to create the first request path
it's fine. When I try and create the second path we get a can't join on
nil error. I debugged this as far as determining it was Rails before
settling on an workaround with more manual path construction.

[1] https://github.com/rails/rails/issues/45554