Refactor ScheduledEditionService out
We’d like to dequeue any prior jobs to ensure there’s only ever one. To do this we’d like to add it into the UpdateService however the controller defers to both the UpdateService and ScheduleService. Instead of adding it to both, we refactor the service layer so UpdateService knows how to publish.