Use a small dispatch object to instruct updates to be now or later
Allow the controller a way to inject a dispatcher object to the service. This tells the service layer how to dispatch this update, now or later.

The controller also wants to share a flash message with the user based on the dispatcher used so we hardcode this in the new object. This could just as easily be passed in. For now it keeps the controller slightly more concise.