Fix loading of GA4 analytics modules
If using the GOV.UK publishing component analytics module loading, the loader does not expect the same structure as the normal GOV.UK module loader. Instead it expects a simple object with an init function to call.

This PR also ensures consistent usage of kebab case for module data attributes