Reduce the number of queries used to build ministers index links
The ministers index presenter was not leveraging eager loading and consequently was running huge numbers of queries. This has the potential to cause performance issues in the Sidekiq queues, as indeed was recently the case with the embassies index presenter.

We also move the govspeak helper from the main presenter to the reshuffle presenter because the main presenter doesn't use any of the methods it provides.