Override the layout navigation
Whitehall has a lot of additional page furniture in the
navigation, which, if we’re linking from Signon, we don’t need, and
could be confusing, so let’s replace it with a slimmed down navigation.

I’ve also added a simple feature test, so we can be alerted if any
changes are made to the layout that result in a change.