Remove overlapping government validation
The content operations team have asked us to remove this validation on the grounds that it makes their workflow difficult when changing the way that governments are labelled.

For example, when changing from labelling governments by election to labelling governments by Prime Minister, it proves awkward to manipulate the dates in such a way that the overlapping constraints are always satisfied.

This risk seems acceptable as only a very small number of users have permission to edit governments and they are responsible for ensuring that the dates are accurate.